+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Support `:fill_color` and `:corner_radius` in `Tucan.bar/4`
 - Support `:fill_color` and `:corner_radius` in `Tucan.histogram/3`
+- Support `:fill_color`, `:point_color` and `:line_color` in `Tucan.area/4`
 
 ## [v0.3.0](https://github.com/pnezis/tucan/tree/v0.3.0) (2024-01-03)